Fixed import of a class since removed in AbstractTripleStore.
BOp#getRequiredProperty(String name) was not compiling under ant due to a generic type parameter for the return type.  I have removed the generic type parameter and added explicit casts everywhere this method is used.

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3552 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/bop/solutions/SliceOp.java
@@ -140,7 +140,7 @@ public class SliceOp extends BindingSetPipelineOp {
      */
     public long getOffset() {
         
-        return getRequiredProperty(Annotations.OFFSET);
+        return (Long) getRequiredProperty(Annotations.OFFSET);
         
     }
 
@@ -149,7 +149,7 @@ public class SliceOp extends BindingSetPipelineOp {
      */
     public long getLimit() {
         
-        return getRequiredProperty(Annotations.LIMIT);
+        return (Long) getRequiredProperty(Annotations.LIMIT);
         
     }
